aiogramarch/templates/project/{{ cookiecutter.project_name }}/requirements.txt
aiogram==3.0.0b6
tortoise-orm==0.19.3
pydantic==1.10.12
pydantic-settings==2.0.2
